# Create closure to do sorting.