--- a/src/core.rs
+++ b/src/core.rs
@@ -118,7 +118,8 @@ impl Core {
     }
 
     fn sort(&self, metas: &mut Vec<Meta>) {
-        metas.sort_unstable_by(|a, b| sort::by_meta(a, b, &self.flags));
+        let sorter = sort::create_sorter(&self.flags);
+        metas.sort_unstable_by(|a, b| (sorter)(a, b));
 
         for meta in metas {
             if let Some(ref mut content) = meta.content {