Was hacking on #277 and realised that splitting the directories out wasn't necessary so jumped to the sorting with a created closure.

## Future architecture for sorting:
1. Take flags and "compose"(not sure how this works in Rust, so I just went with a loop), sorting functions together to create the sorting function
1. Cache sorting function to reuse for the recursive calls to `Core#sort`

The test pass but I would like to do some performance testing and complete the second step.