@aldhsu 2019-09-13
# Performance testing
Using the closure:
```
~/Code/lsd/target/release/lsd -R --group-dirs=last  1.93s user 1.47s system 89% cpu 3.802 total
```

Without the closure:
```
lsd -R --group-dirs=last  1.91s user 1.49s system 89% cpu 3.821 total
```

Doesn't look like much has changed.