@aldhsu 2019-12-17
Branch is up to date now.
In the current state, it should be simpler to add new sort criteria,  add support for specifying sort direction on individual fields and multisorting.

Possibly minor effect on performance, recursive sorting around 100k files:
Before
```
Benchmark #1: lsd --recursive -r --sizesort
  Time (mean ± σ):      2.171 s ±  0.015 s    [User: 1.108 s, System: 1.037 s]
  Range (min … max):    2.149 s …  2.187 s    10 runs
```
After, branch built with --release
```
Benchmark #1: /Users/allen.hsu/Code/lsd/target/release/lsd --recursive -r --sizesort
  Time (mean ± σ):      2.261 s ±  0.032 s    [User: 1.187 s, System: 1.042 s]
  Range (min … max):    2.218 s …  2.328 s    10 runs
```