@aldhsu 2020-08-17
@zwpaper I've updated the PR. I have a bit more knowledge about Rust now and I realized the closure is actually not necessary since closures are just implicit structs. I switched it to store a vec of function pointers on the Core struct instead of a closure, I think this is a bit easier to work with.

# Before
```
Benchmark #1: lsd -R --group-dirs=last
  Time (mean ± σ):     885.4 ms ±  26.4 ms    [User: 551.8 ms, System: 318.1 ms]
  Range (min … max):   854.0 ms … 938.9 ms    10 runs
```

# After
```
Benchmark #1: ~/Code/lsd/target/release/lsd -R --group-dirs=last
  Time (mean ± σ):     923.2 ms ±  23.5 ms    [User: 596.3 ms, System: 318.9 ms]
  Range (min … max):   909.3 ms … 988.2 ms    10 runs
```
This still definitely makes sorting a bit slower.